pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "he2plus"
version = "0.1.0"
authors = [
    {name = "Prakhar Tripathi", email = "prakhar@he2plus.dev"},
]
description = "A powerful Python library for developer productivity and automation - built by a dev frustrated by dependency issues"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["development", "automation", "productivity", "setup", "dependencies", "ml", "cloud", "web3"]
dependencies = [
    "requests>=2.28.0",
    "psutil>=5.9.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "colorama>=0.4.0",
    "rich>=13.0.0",
    "inquirer>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]
ml = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tensorflow>=2.8.0",
    "torch>=1.11.0",
    "jupyter>=1.0.0",
]
cloud = [
    "boto3>=1.21.0",
    "azure-sdk>=1.0.0",
    "google-cloud>=0.34.0",
    "docker>=5.0.0",
    "kubernetes>=18.0.0",
]
web3 = [
    "web3>=5.0.0",
    "eth-brownie>=1.18.0",
    "hardhat>=0.1.0",
    "solana>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/he2plus/he2plus"
Documentation = "https://github.com/he2plus/he2plus#readme"
Repository = "https://github.com/he2plus/he2plus"
"Bug Tracker" = "https://github.com/he2plus/he2plus/issues"
"Twitter" = "https://twitter.com/he2plus"

[project.scripts]
he2plus = "he2plus.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
he2plus = ["*.txt", "*.md", "*.yaml", "*.json"]

[tool.setuptools.packages.find]
where = ["."]
include = ["he2plus*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
